LPS-44579 Create the exception using service.xml and move it to portal-service with the rest of the exceptions
