+++ portal-impl/src/com/liferay/portlet/blogs/service.xml
@@ -233,5 +233,6 @@
 		<exception>EntrySmallImageName</exception>
 		<exception>EntrySmallImageSize</exception>
 		<exception>EntryTitle</exception>
+		<exception>TrackbackValidation</exception>
 	</exceptions>
 </service-builder>
